RF | UserConfigPage extends BaseConfigPage
