+++ src/test/java/school/redrover/model/StatusUserPage.java
@@ -55,10 +55,10 @@ public class StatusUserPage extends BaseMainHeaderPage<StatusUserPage> {
                 By.xpath("//div[@id='main-panel']/div[contains(text(), 'ID')]")).getText();
     }
 
-    public ConfigureUserPage clickConfigureSideMenu() {
+    public UserConfigPage clickConfigureSideMenu() {
         getWait5().until(ExpectedConditions.elementToBeClickable(
                 getDriver().findElement(By.cssSelector("[href$='/configure']")))).click();
 
-        return new ConfigureUserPage(getDriver());
+        return new UserConfigPage(new StatusUserPage(getDriver()));
     }
 }
